+++ pages/components/style/LoadDetail.module.css
