+++ pages/components/style/LoadingSpinner.module.css
