+++ pages/components/style/Search.module.css
