+++ pages/components/style/Trending.module.css
