+++ pages/layouts/NavLayout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import Nav from "@/components/main/nav/Nav";
+import Nav from "@/pages/components/main/nav/Nav";
 
 export default function Layout({ children }: { children: ReactNode }) {
   return (
