+++ pages/types/testFiles/NewsItemCheckType.test.ts
