+++ pages/types/testFiles/NewsType.test.ts
