+++ pages/types/typeFiles/ArticleType.ts
