+++ pages/types/typeFiles/CategoryType.ts
