+++ pages/types/typeFiles/NewsItemCheckType.ts
