+++ pages/types/typeFiles/NewsType.ts
