+++ pages/utility/newsHandleNavigation.ts
